requirements.txt
torch       # According to CUDA Version
torchvision # According to CUDA Version
timm
pillow
jupyter
flask # For Frontend
# For Cuda Installation:
# 1. Download Cuda Toolkit from Nvidia and Install.
# 2. Download cuDNN Zip File. Extract the Zip and Paste the Files to the Cuda Toolkit Folder According to the Directories
# 3. pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118 (Install Newer Version if Required.)